feat(board): Write method to simplify fen string
by simplify, I mean it takes a complete fen string and returns only the
part that's needed to arrange the pieces. It also changes something like
'3pp2' to '111pp11' which helps give each rank in the fen string equal
length. This is done because it would be so much trickier trying to match
the the characters in the fen string to each square on the board if the
indexes didn't match.