fix(position): Let Position create invalid positions
a feature was added previously to stop position objects from generating
invalid positions, but it'll lead to issues especially for the knight,
because depending on where it is on the board, some incorrect movements
will be made. Now, the feature has been removed, so that the list of
positions will be cleaned later and the invalid ones, removed.