feat(pawn): Add #white_pawn_moves for normal + attack moves
the reason I took this approach is, at the moment, moves for all pieces
goes through a process where positions that include enemy pieces are
retained because the pieces attack any enemies in their normal path.
but there's a difference when it comes to pawns. Pawns do not attack
enemies that are in their normal forward paths, they instead attack in
diagonal directions. So, to overcome this, I'll modify the pipeline for
the pawns to remove any positions with pieces on there regardless of
color. This way, pawns will not proceed if there are pieces in the way.
However, it creates another complication for attack moves, because those
are positions with enemiy pieces on them and we don't want those removed.
Hence this approach to route the attack positions separately with the
aim of adding them back in after the weeding out has occured.